# JSON as fallback encoding