**What this PR does / why we need it**:

The output of `Sprintf("%+v")` is neither readable nor parseable. JSON encoding
is better in both regards.

Because JSON follows pointers instead of printing their value, output becomes more useful, too.
 
There are some downsides:
- Unexported fields don't get encoded anymore. However, the same happens
   with other backends (in particular, zapr), so one could also
   argue that making the text format behave like those others
   is good because developers notice early that they need custom
   String and MarshalLog methods if they want to log such fields.
- The result KObjs (a slice) now gets encoded as array of structs.
   KObjs already got deprecated a year ago. KObjSlice should be used
   instead.

**Which issue(s) this PR fixes**:

Fixes #https://github.com/kubernetes/klog/issues/352

**Special notes for your reviewer**:

The first commit introduces some new output unit tests, the second has the actual change.

**Release note**:
```release-note
Structured logging now uses JSON encoding for structs, slices and maps. For pointers embedded in those, the thing they point to gets printed, not the pointer values. Unexported fields in structs no longer get included in the output. Implement `String` and `MarshalLog` for structs if such fields are important.
```