@pohly 2023-04-26
In most cases, performance is better than before (-17.55% for structs, -11.69% for slices). Printing maps becomes a bit slower, but not significantly (6.04µs ± 8%  ->  6.14µs ±14%).

<details>

```
name                                                                            old time/op    new time/op    delta
pkg:k8s.io/klog/v2 goos:linux goarch:amd64
Header-36                                                                         1.63µs ± 9%    1.72µs ± 3%     ~     (p=0.151 n=5+5)
HeaderWithDir-36                                                                  1.63µs ±10%    1.69µs ± 5%     ~     (p=0.341 n=5+5)
V-36                                                                              2.60ns ± 0%    2.37ns ± 0%   -8.97%  (p=0.008 n=5+5)
KRef-36                                                                           0.99ns ± 6%    0.99ns ± 6%     ~     (p=0.841 n=5+5)
KObj-36                                                                           6.22ns ± 0%    6.23ns ± 0%     ~     (p=0.167 n=5+5)
KObjs/0/simple-36                                                                  280ns ± 3%     281ns ± 6%     ~     (p=0.548 n=5+5)
KObjs/0/conditional-36                                                            2.61ns ± 0%    2.61ns ± 0%     ~     (p=0.579 n=5+5)
KObjs/10/simple-36                                                                1.17µs ± 4%    1.18µs ± 6%     ~     (p=0.841 n=5+5)
KObjs/10/conditional-36                                                           2.61ns ± 0%    2.61ns ± 0%     ~     (p=0.143 n=5+5)
KObjs/20/simple-36                                                                1.93µs ± 1%    1.93µs ± 2%     ~     (p=1.000 n=5+5)
KObjs/20/conditional-36                                                           2.61ns ± 1%    2.60ns ± 0%     ~     (p=0.270 n=5+4)
KObjs/30/simple-36                                                                2.62µs ± 2%    2.63µs ± 1%     ~     (p=0.937 n=5+5)
KObjs/30/conditional-36                                                           2.60ns ± 0%    2.61ns ± 0%     ~     (p=0.571 n=5+5)
KObjs/40/simple-36                                                                3.33µs ± 1%    3.34µs ± 1%     ~     (p=1.000 n=5+5)
KObjs/40/conditional-36                                                           2.61ns ± 0%    2.61ns ± 0%     ~     (p=0.421 n=5+5)
KObjs/50/simple-36                                                                4.06µs ± 2%    4.05µs ± 1%     ~     (p=0.881 n=5+5)
KObjs/50/conditional-36                                                           2.61ns ± 0%    2.61ns ± 0%   -0.32%  (p=0.040 n=5+5)
KObjs/60/simple-36                                                                5.02µs ± 5%    5.14µs ± 2%     ~     (p=0.254 n=5+5)
KObjs/60/conditional-36                                                           2.61ns ± 0%    2.60ns ± 1%     ~     (p=0.230 n=5+5)
KObjs/70/simple-36                                                                5.60µs ± 3%    5.59µs ± 4%     ~     (p=1.000 n=5+5)
KObjs/70/conditional-36                                                           2.61ns ± 0%    2.61ns ± 0%     ~     (p=0.857 n=5+5)
KObjs/80/simple-36                                                                6.62µs ± 3%    6.38µs ± 4%     ~     (p=0.056 n=5+5)
KObjs/80/conditional-36                                                           2.61ns ± 0%    2.61ns ± 0%     ~     (p=1.000 n=5+5)
KObjs/90/simple-36                                                                6.90µs ± 4%    7.01µs ± 1%     ~     (p=0.222 n=5+5)
KObjs/90/conditional-36                                                           2.61ns ± 0%    2.61ns ± 1%     ~     (p=0.389 n=5+5)
KObjs/100/simple-36                                                               7.83µs ± 4%    7.66µs ± 2%     ~     (p=0.222 n=5+5)
KObjs/100/conditional-36                                                          2.61ns ± 0%    2.60ns ± 0%     ~     (p=0.056 n=5+5)
KObjSlice/0/simple-36                                                              264ns ± 6%     262ns ± 4%     ~     (p=1.000 n=5+5)
KObjSlice/0/conditional-36                                                        2.61ns ± 0%    2.60ns ± 0%     ~     (p=0.492 n=5+4)
KObjSlice/10/simple-36                                                             263ns ± 4%     261ns ± 3%     ~     (p=0.548 n=5+5)
KObjSlice/10/conditional-36                                                       2.61ns ± 1%    2.60ns ± 0%     ~     (p=0.310 n=5+5)
KObjSlice/20/simple-36                                                             265ns ± 8%     259ns ± 6%     ~     (p=0.548 n=5+5)
KObjSlice/20/conditional-36                                                       2.61ns ± 0%    2.61ns ± 0%     ~     (p=0.373 n=5+5)
KObjSlice/30/simple-36                                                             267ns ± 6%     259ns ± 7%     ~     (p=0.421 n=5+5)
KObjSlice/30/conditional-36                                                       2.61ns ± 0%    2.61ns ± 0%     ~     (p=0.786 n=5+5)
KObjSlice/40/simple-36                                                             267ns ± 3%     264ns ± 4%     ~     (p=0.690 n=5+5)
KObjSlice/40/conditional-36                                                       2.61ns ± 1%    2.60ns ± 0%     ~     (p=0.198 n=5+5)
KObjSlice/50/simple-36                                                             266ns ± 2%     264ns ± 8%     ~     (p=0.794 n=5+5)
KObjSlice/50/conditional-36                                                       2.60ns ± 0%    2.61ns ± 0%     ~     (p=0.151 n=5+5)
KObjSlice/60/simple-36                                                             265ns ± 6%     263ns ± 7%     ~     (p=1.000 n=5+5)
KObjSlice/60/conditional-36                                                       2.61ns ± 1%    2.60ns ± 0%   -0.39%  (p=0.024 n=5+5)
KObjSlice/70/simple-36                                                             262ns ±10%     271ns ± 2%     ~     (p=0.206 n=5+5)
KObjSlice/70/conditional-36                                                       2.61ns ± 0%    2.61ns ± 0%     ~     (p=0.841 n=5+5)
KObjSlice/80/simple-36                                                             262ns ± 4%     265ns ± 3%     ~     (p=0.690 n=5+5)
KObjSlice/80/conditional-36                                                       2.61ns ± 0%    2.60ns ± 1%     ~     (p=0.063 n=4+5)
KObjSlice/90/simple-36                                                             270ns ± 3%     263ns ± 3%     ~     (p=0.222 n=5+5)
KObjSlice/90/conditional-36                                                       2.61ns ± 0%    2.60ns ± 0%     ~     (p=0.135 n=5+5)
KObjSlice/100/simple-36                                                            267ns ± 9%     267ns ± 3%     ~     (p=0.548 n=5+5)
KObjSlice/100/conditional-36                                                      2.60ns ± 0%    2.61ns ± 0%     ~     (p=0.984 n=5+5)
Scalars/simple-36                                                                  111ns ± 2%     112ns ± 3%     ~     (p=0.548 n=5+5)
Scalars/conditional-36                                                            2.62ns ± 0%    2.64ns ± 1%     ~     (p=0.683 n=5+5)
ScalarsWithLogger/simple-36                                                        119ns ± 3%     117ns ± 3%     ~     (p=0.222 n=5+5)
ScalarsWithLogger/conditional-36                                                  4.25ns ± 0%    3.89ns ± 0%   -8.30%  (p=0.008 n=5+5)
KObjSliceWithLogger/simple-36                                                      278ns ± 6%     276ns ± 4%     ~     (p=0.421 n=5+5)
KObjSliceWithLogger/conditional-36                                                4.25ns ± 0%    4.03ns ± 0%   -5.25%  (p=0.008 n=5+5)
Logs-36                                                                           7.12µs ± 4%    7.19µs ± 7%     ~     (p=0.548 n=5+5)
WithoutDeferUnLock-36                                                             11.9ns ± 0%    12.0ns ± 2%     ~     (p=0.325 n=5+5)
WithDeferUnLock-36                                                                13.3ns ± 0%    13.3ns ± 0%     ~     (p=0.127 n=5+5)
KlogOutput/KObjSlice_int_arg-36                                                   3.62µs ± 5%    3.59µs ± 4%     ~     (p=0.841 n=5+5)
KlogOutput/handle_integer_keys-36                                                 5.89µs ± 4%    5.83µs ± 3%     ~     (p=0.690 n=5+5)
KlogOutput/handle_odd-numbers_of_KVs_in_both_log_values_and_Info_args-36          5.94µs ± 4%    5.95µs ± 9%     ~     (p=0.690 n=5+5)
KlogOutput/vmodule-36                                                             4.92µs ±11%    4.84µs ± 8%     ~     (p=1.000 n=5+5)
KlogOutput/other_vmodule-36                                                        845ns ± 0%     848ns ± 0%     ~     (p=0.095 n=5+5)
KlogOutput/override_WithValues-36                                                 4.57µs ± 6%    4.33µs ± 4%   -5.25%  (p=0.032 n=5+5)
KlogOutput/print_duplicate_keys_in_arguments-36                                   3.88µs ±10%    3.98µs ± 6%     ~     (p=0.690 n=5+5)
KlogOutput/map_keys-36                                                            4.52µs ± 5%    4.77µs ± 6%     ~     (p=0.151 n=5+5)
KlogOutput/verbosity_disabled-36                                                  15.2ns ± 0%    15.2ns ± 0%     ~     (p=0.103 n=5+5)
KlogOutput/handle_odd-numbers_of_KVs-36                                           3.95µs ± 8%    4.03µs ± 3%     ~     (p=0.841 n=5+5)
KlogOutput/html_characters-36                                                     3.56µs ± 7%    3.48µs ± 9%     ~     (p=0.421 n=5+5)
KlogOutput/struct_keys-36                                                         5.18µs ± 4%    5.48µs ± 3%   +5.76%  (p=0.024 n=5+5)
KlogOutput/slice_values-36                                                        4.50µs ± 4%    3.97µs ± 4%  -11.69%  (p=0.008 n=5+5)
KlogOutput/call_depth-36                                                          3.79µs ± 7%    3.70µs ± 2%     ~     (p=0.690 n=5+5)
KlogOutput/KObjSlice_nil_entry-36                                                 3.52µs ±11%    3.59µs ± 9%     ~     (p=0.690 n=5+5)
KlogOutput/String()_for_nil-36                                                    7.80µs ± 7%    7.93µs ± 9%     ~     (p=0.690 n=5+5)
KlogOutput/KObjSlice_okay-36                                                      3.65µs ± 2%    3.77µs ±12%     ~     (p=0.151 n=5+5)
KlogOutput/odd_WithValues-36                                                      12.9µs ± 5%    13.4µs ± 7%     ~     (p=0.222 n=5+5)
KlogOutput/multiple_WithValues-36                                                 16.2µs ± 4%    16.7µs ± 6%     ~     (p=0.310 n=5+5)
KlogOutput/preserve_order_of_key/value_pairs-36                                   7.02µs ± 5%    6.72µs ± 7%     ~     (p=0.095 n=5+5)
KlogOutput/KObjSlice_ints-36                                                      3.81µs ±12%    3.79µs ± 6%     ~     (p=0.841 n=5+5)
KlogOutput/regular_error_types_as_value-36                                        3.54µs ± 2%    3.40µs ± 7%     ~     (p=0.421 n=5+5)
KlogOutput/ignore_MarshalJSON-36                                                  3.51µs ± 3%    3.57µs ± 5%     ~     (p=0.460 n=5+5)
KlogOutput/String()_that_panics-36                                                4.54µs ± 4%    4.55µs ± 3%     ~     (p=1.000 n=5+5)
KlogOutput/MarshalLog()_that_panics-36                                            4.71µs ± 1%    4.69µs ± 3%     ~     (p=0.690 n=5+5)
KlogOutput/Error()_for_nil-36                                                     7.24µs ± 3%    7.94µs ± 3%   +9.70%  (p=0.008 n=5+5)
KlogOutput/log_with_values-36                                                     3.47µs ± 7%    3.52µs ± 3%     ~     (p=0.421 n=5+5)
KlogOutput/override_single_value-36                                               3.40µs ± 7%    3.65µs ± 4%   +7.58%  (p=0.032 n=5+5)
KlogOutput/KObjs-36                                                               4.78µs ± 3%    4.05µs ± 6%  -15.32%  (p=0.008 n=5+5)
KlogOutput/KObjSlice_nil_arg-36                                                   3.25µs ± 8%    3.32µs ± 5%     ~     (p=0.841 n=5+5)
KlogOutput/regular_error_types_when_using_logr.Error-36                           3.61µs ± 2%    3.45µs ± 4%   -4.53%  (p=0.032 n=5+5)
KlogOutput/MarshalLog()_for_nil-36                                                4.97µs ± 7%    4.89µs ± 2%     ~     (p=0.690 n=5+5)
KlogOutput/Error()_that_panics-36                                                 4.48µs ± 5%    4.32µs ± 4%     ~     (p=0.095 n=5+5)
KlogOutput/verbosity_enabled-36                                                   3.29µs ± 9%    3.26µs ± 1%     ~     (p=0.730 n=5+4)
KlogOutput/struct_values-36                                                       4.48µs ± 5%    3.69µs ± 5%  -17.55%  (p=0.008 n=5+5)
KlogOutput/quotation-36                                                           3.91µs ± 6%    3.77µs ± 3%     ~     (p=0.151 n=5+5)
KlogOutput/KObj-36                                                                3.44µs ± 2%    3.52µs ± 4%     ~     (p=0.151 n=5+5)
KlogOutput/MarshalLog()_that_returns_itself-36                                    3.77µs ± 9%    3.70µs ± 7%     ~     (p=0.690 n=5+5)
KlogOutput/empty_WithValues-36                                                    3.08µs ± 8%    3.08µs ± 5%     ~     (p=0.690 n=5+5)
KlogOutput/log_with_multiple_names_and_values-36                                  3.90µs ± 6%    4.07µs ± 3%     ~     (p=0.095 n=5+5)
KlogOutput/map_values-36                                                          6.04µs ± 8%    6.14µs ±14%     ~     (p=0.548 n=5+5)
KlogOutput/log_with_name_and_values-36                                            3.63µs ±10%    3.71µs ± 7%     ~     (p=0.841 n=5+5)
pkg:k8s.io/klog/v2/internal/benchmarks/contextvalues goos:linux goarch:amd64
NewContext1Percent-36                                                             61.7µs ± 3%    63.5µs ± 2%   +2.88%  (p=0.008 n=5+5)
NewContext100PercentDisabled-36                                                   78.6µs ± 2%    80.4µs ± 4%     ~     (p=0.151 n=5+5)
NewContext100Percent-36                                                           82.6µs ± 1%    82.7µs ± 2%     ~     (p=1.000 n=5+5)
NewContext1PercentValues-36                                                       25.2µs ± 2%    25.7µs ± 2%   +2.19%  (p=0.048 n=5+5)
NewContext100PercentDisabledValues-36                                             41.7µs ± 3%    43.0µs ± 1%   +3.02%  (p=0.016 n=5+5)
NewContext100PercentValues-36                                                     43.9µs ± 2%    45.8µs ± 1%   +4.24%  (p=0.008 n=5+5)
pkg:k8s.io/klog/v2/internal/serialize goos:linux goarch:amd64
MergeKVs/0x0/no-duplicates-36                                                     2.14ns ± 0%    2.13ns ± 0%     ~     (p=0.087 n=5+5)
MergeKVs/0x0/all-duplicates-36                                                    2.13ns ± 1%    2.13ns ± 0%     ~     (p=0.603 n=5+5)
MergeKVs/0x0/start-duplicate-36                                                    273ns ± 4%     282ns ± 5%     ~     (p=0.222 n=5+5)
MergeKVs/0x0/end-duplicate-36                                                      275ns ± 5%     276ns ± 5%     ~     (p=0.690 n=5+5)
MergeKVs/0x1/no-duplicates-36                                                     2.61ns ± 0%    2.61ns ± 0%     ~     (p=1.000 n=5+5)
MergeKVs/0x1/all-duplicates-36                                                    2.61ns ± 0%    2.61ns ± 0%     ~     (p=1.000 n=5+5)
MergeKVs/0x1/start-duplicate-36                                                    368ns ± 3%     375ns ± 5%     ~     (p=0.151 n=5+5)
MergeKVs/0x1/end-duplicate-36                                                      372ns ± 8%     367ns ± 4%     ~     (p=0.841 n=5+5)
MergeKVs/0x2/no-duplicates-36                                                     2.61ns ± 0%    2.61ns ± 0%     ~     (p=0.817 n=5+5)
MergeKVs/0x2/all-duplicates-36                                                    2.60ns ± 1%    2.61ns ± 0%     ~     (p=0.389 n=5+5)
MergeKVs/0x2/start-duplicate-36                                                    439ns ± 5%     438ns ± 3%     ~     (p=1.000 n=5+5)
MergeKVs/0x2/end-duplicate-36                                                      449ns ± 2%     451ns ± 2%     ~     (p=0.841 n=5+5)
MergeKVs/0x3/no-duplicates-36                                                     2.60ns ± 1%    2.61ns ± 0%     ~     (p=0.881 n=5+5)
MergeKVs/0x3/all-duplicates-36                                                    2.61ns ± 0%    2.61ns ± 0%     ~     (p=0.810 n=5+5)
MergeKVs/0x3/start-duplicate-36                                                    553ns ± 5%     540ns ± 6%     ~     (p=0.421 n=5+5)
MergeKVs/0x3/end-duplicate-36                                                      551ns ± 3%     556ns ± 3%     ~     (p=0.841 n=5+5)
MergeKVs/0x4/no-duplicates-36                                                     2.61ns ± 0%    2.60ns ± 0%     ~     (p=0.143 n=5+5)
MergeKVs/0x4/all-duplicates-36                                                    2.61ns ± 0%    2.60ns ± 1%     ~     (p=0.540 n=5+5)
MergeKVs/0x4/start-duplicate-36                                                    641ns ± 4%     634ns ± 3%     ~     (p=0.381 n=5+5)
MergeKVs/0x4/end-duplicate-36                                                      650ns ± 2%     622ns ± 5%     ~     (p=0.056 n=5+5)
MergeKVs/0x5/no-duplicates-36                                                     2.61ns ± 0%    2.60ns ± 1%     ~     (p=0.722 n=5+5)
MergeKVs/0x5/all-duplicates-36                                                    2.61ns ± 0%    2.61ns ± 0%     ~     (p=0.833 n=5+5)
MergeKVs/0x5/start-duplicate-36                                                    734ns ± 3%     729ns ± 3%     ~     (p=0.690 n=5+5)
MergeKVs/0x5/end-duplicate-36                                                      733ns ± 8%     747ns ± 5%     ~     (p=0.690 n=5+5)
MergeKVs/0x6/no-duplicates-36                                                     2.61ns ± 1%    2.61ns ± 0%     ~     (p=0.675 n=5+5)
MergeKVs/0x6/all-duplicates-36                                                    2.61ns ± 0%    2.61ns ± 0%     ~     (p=0.063 n=4+5)
MergeKVs/0x6/start-duplicate-36                                                    795ns ± 4%     784ns ± 4%     ~     (p=0.333 n=5+5)
MergeKVs/0x6/end-duplicate-36                                                      806ns ± 3%     815ns ± 4%     ~     (p=0.690 n=5+5)
MergeKVs/0x7/no-duplicates-36                                                     2.61ns ± 0%    2.61ns ± 0%     ~     (p=1.000 n=5+5)
MergeKVs/0x7/all-duplicates-36                                                    2.61ns ± 1%    2.61ns ± 0%     ~     (p=0.413 n=5+5)
MergeKVs/0x7/start-duplicate-36                                                    924ns ± 3%     899ns ± 3%     ~     (p=0.222 n=5+5)
MergeKVs/0x7/end-duplicate-36                                                      941ns ± 2%     930ns ± 5%     ~     (p=0.690 n=5+5)
MergeKVs/0x8/no-duplicates-36                                                     2.60ns ± 0%    2.61ns ± 0%     ~     (p=1.000 n=5+5)
MergeKVs/0x8/all-duplicates-36                                                    2.61ns ± 0%    2.61ns ± 0%     ~     (p=0.889 n=5+5)
MergeKVs/0x8/start-duplicate-36                                                   1.93µs ± 2%    1.95µs ± 4%     ~     (p=0.841 n=5+5)
MergeKVs/0x8/end-duplicate-36                                                     1.98µs ± 4%    1.98µs ± 5%     ~     (p=1.000 n=5+5)
MergeKVs/0x9/no-duplicates-36                                                     2.61ns ± 0%    2.61ns ± 0%     ~     (p=1.000 n=5+5)
…